+++ wuk/includes/utils/number.hh
@@ -15,7 +15,7 @@ namespace wuk::utils {
         wuk::Buffer result{sizeof(T), 0};
 
         T swap = (!reverse_endianness) ? val : swap_endian(val);
-        memcpy(result.write(sizeof(T)), &swap, sizeof(T));
+        memcpy(result.write<wuk::byte>(sizeof(T)), &swap, sizeof(T));
 
         return result;
     }
@@ -27,7 +27,6 @@ namespace wuk::utils {
             std::is_integral_v<T> || std::is_floating_point_v<T>,
             "wuk::utils::unpack_bytes requires numeric type."
         );
-        // 注意，没有任何问题，就是缓冲区的内存大小
         if (buffer.capacity() < sizeof(T)) {
             throw wuk::Exception(wuk::Error::MEMORY, "wuk::utils::unpack_bytes",
                 "The buffer is too small, please adjust the size.");